Moved HtmlStripCharFilterFactory to analyis.common package like the other factories.
