+++ modules/analysis-common/src/main/java/org/elasticsearch/analysis/common/HtmlStripCharFilterFactory.java
@@ -17,12 +17,13 @@
  * under the License.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.analysis.common;
 
 import org.apache.lucene.analysis.charfilter.HTMLStripCharFilter;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.IndexSettings;
+import org.elasticsearch.index.analysis.AbstractCharFilterFactory;
 
 import java.io.Reader;
 import java.util.Set;
@@ -33,7 +34,7 @@ import static org.elasticsearch.common.util.set.Sets.newHashSet;
 public class HtmlStripCharFilterFactory extends AbstractCharFilterFactory {
     private final Set<String> escapedTags;
 
-    public HtmlStripCharFilterFactory(IndexSettings indexSettings, Environment env, String name, Settings settings) {
+    HtmlStripCharFilterFactory(IndexSettings indexSettings, Environment env, String name, Settings settings) {
         super(indexSettings, name);
         String[] escapedTags = settings.getAsArray("escaped_tags");
         if (escapedTags.length > 0) {
@@ -41,10 +42,6 @@ public class HtmlStripCharFilterFactory extends AbstractCharFilterFactory {
         } else {
             this.escapedTags = null;
         }
-    }
-
-    public Set<String> escapedTags() {
-        return escapedTags;
     }
 
     @Override
